requirements.txt
# 基础依赖
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
playwright==1.40.0

# 数据处理和存储
pandas==2.1.3
sqlalchemy==2.0.23
pymongo==4.6.0

# AI和分析
scikit-learn==1.3.2
transformers==4.35.2
nltk==3.8.1

# 工具和调度
python-dotenv==1.0.0
apscheduler==3.10.4
loguru==0.7.2 